api/requirements.txt
fastapi==0.115.12
fastapi-cli==0.0.7
python-dotenv==1.1.0
pydantic==2.11.5
pydantic_core==2.33.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10